Split extract TranslationService from TranslateHelper
